+++ src/main/kotlin/io/gitlab/arturbosch/detekt/idea/intention/AutoCorrectIntention.kt
@@ -1,13 +1,14 @@
 package io.gitlab.arturbosch.detekt.idea.intention
 
 import com.intellij.codeInsight.intention.IntentionAction
+import com.intellij.codeInsight.intention.LowPriorityAction
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
 import io.gitlab.arturbosch.detekt.idea.DETEKT
 import io.gitlab.arturbosch.detekt.idea.action.AutoCorrectAction
 
-class AutoCorrectIntention : IntentionAction {
+class AutoCorrectIntention : IntentionAction, LowPriorityAction {
 
     override fun startInWriteAction(): Boolean = true